Added files from the Workstation at Calavera
The files included are of similar format but range with type. There are enrichment fliers, gym schedules, and internal use documents I have created for staff. The "Parks and Employee Handbook Cover" is slightly modified from the original because I did not like how MS Word added the gray background to the logo when that particular style was used.