Added Park layout documents
The attached files are the images and documents I created for internal use for Specialist staff to successfully show patrons which areas are up for reservation at our sites. The Alga Norte one was never completed due to complexity in the area and my lack of knowledge for the park layout.